transform each element face
Previously each element face entity's center was the center of the
block, and I set triangle mesh coordinates in the appropriate positions
relative to that center. This change puts the center of each face in the
center of its mesh, and translates the mesh into position. This seems to
reduce visual glitches from alpha blending ordering.